feat: created a new header and made the app component cleaner
